Update commonKnowledgeCOLOR.md
Temporary table fix, made the 100+ entries of the table collapsible. Thinking i might section the values with collapsing sections so it isn't nearly as long. For anyone wondering why I want to have a table instead of an image, is so that people can copy and paste values. Just a small quality of life thing that should be used.